+++ server/models/login.models.go
@@ -10,7 +10,7 @@ type ClientCredentials struct {
 }
 
 type RestaurantCredentials struct {
-	Id       int    `json:"id"`
+	Id       uint64 `json:"id"`
 	Password string `json:"password"`
 }
 
@@ -21,12 +21,12 @@ type Response struct {
 
 type ClientClaims struct {
 	Email string `json:"email"`
-	Id    int    `json:"id"`
+	Id    uint64 `json:"id"`
 	Role  string `json:"role"`
 	jwt.StandardClaims
 }
 
 type RestaurantClaims struct {
-	Id int `json:"id"`
+	Id uint64 `json:"id"`
 	jwt.StandardClaims
 }
